corrected

+++ src/scripts/main.js
@@ -6,44 +6,44 @@ const firstRow = table.rows[0];
 
 buttons.forEach((button) => {
   button.addEventListener('click', (ev) => {
-    const buttonClass = button.classList[0];
+    const buttonClass = button.classList;
 
-    switch (buttonClass) {
-      case 'append-row':
-        if (table.rows.length < 10) {
-          const newRow = table.insertRow();
+    if (buttonClass.contains('append-row')) {
+      if (table.rows.length < 10) {
+        const newRow = table.insertRow();
 
-          for (let cell = 0; cell < firstRow.cells.length; cell++) {
-            const newCell = newRow.insertCell(cell);
+        for (let cell = 0; cell < firstRow.cells.length; cell++) {
+          const newCell = newRow.insertCell(cell);
 
-            newCell.textContent = '';
-          }
+          newCell.textContent = '';
         }
+      }
+    }
 
-        break;
-      case 'remove-row':
-        if (table.rows.length > 2) {
-          table.deleteRow(-1);
-        }
-        break;
-      case 'append-column':
-        if (table.rows[0].cells.length < 10) {
-          const rows = document.querySelectorAll('tr');
+    if (buttonClass.contains('remove-row')) {
+      if (table.rows.length > 2) {
+        table.deleteRow(-1);
+      }
+    }
 
-          rows.forEach((curRow) => {
-            const newCell = curRow.insertCell(-1);
+    if (buttonClass.contains('append-column')) {
+      if (table.rows[0].cells.length < 10) {
+        const rows = document.querySelectorAll('tr');
 
-            newCell.textContent = '';
-          });
+        rows.forEach((curRow) => {
+          const newCell = curRow.insertCell(-1);
+
+          newCell.textContent = '';
+        });
+      }
+    }
+
+    if (buttonClass.contains('remove-column')) {
+      if (firstRow.cells.length > 2) {
+        for (const curRow of table.rows) {
+          curRow.deleteCell(-1);
         }
-        break;
-      case 'remove-column':
-        if (firstRow.cells.length > 2) {
-          for (const curRow of table.rows) {
-            curRow.deleteCell(-1);
-          }
-        }
-        break;
+      }
     }
   });
 });
